splash screen file renamed
